+++ Example/NavigationDemo/Common/View/BottomSheet/BottomSheetHeightProvider.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+protocol BottomSheetHeightProvider: AnyObject {
+    func bottomSheetHeight(forContainerSize containerSize: CGSize) -> CGFloat
+}
+
+final class DefaultBottomSheetHeightProvider: BottomSheetHeightProvider {
+    func bottomSheetHeight(forContainerSize containerSize: CGSize) -> CGFloat {
+        return containerSize.height * 0.4
+    }
+}