+++ Example/NavigationDemo/Common/View/BottomSheet/BottomSheetPresentableLifecycle.swift
@@ -1,0 +1,9 @@
+protocol BottomSheetPresentableLifecycle: AnyObject {
+    func bottomSheetDidChangeViewState(state: BottomSheetViewState)
+    func bottomSheetShouldDismiss() -> Bool
+}
+
+extension BottomSheetPresentableLifecycle {
+    func bottomSheetShouldDismiss() -> Bool { true }
+    func bottomSheetDidChangeViewState(state: BottomSheetViewState) { }
+}