+++ Example/NavigationDemo/VIPER/Shelf/Assembly/ShelfAssembly.swift
@@ -1,7 +1,16 @@
 import UIKit
 import Marshroute
 
+enum ShelfStyle: Int {
+    case root
+    case modalWithNavigationBar
+    case modalBottomSheetWithNavigationBar
+    case modalBottomSheetWithoutNavigationBar
+}
+
 protocol ShelfAssembly: AnyObject {
-    func module(routerSeed: RouterSeed)
-        -> UIViewController
+    func module(
+        style: ShelfStyle,
+        routerSeed: RouterSeed)
+        -> (ShelfModule, UIViewController)
 }