+++ Example/NavigationDemo/VIPER/Shelf/Presenter/ShelfModuleInput.swift
@@ -1,0 +1,8 @@
+enum ShelfResult {
+    case cancelled
+}
+
+protocol ShelfModule: AnyObject {
+    func dismissCurrentModule()
+    var onFinish: ((ShelfModule, ShelfResult) -> ())? { get set }
+}