+++ Marshroute/Sources/Transitions/TransitionAnimationsLaunching/ModalMasterDetail/Presentation/ModalMasterDetailPresentationAnimationLaunchingContext.swift
@@ -26,11 +26,7 @@ public struct ModalMasterDetailPresentationAnimationLaunchingContext {
         }
         
         if sourceViewController?.presentedViewController == nil {
-            marshrouteAssertionFailure(
-                """
-                It looks like \(targetViewController as Any) did not deallocate due to some retain cycle! 
-                """
-            )
+            assertPossibleRetainCycle(ofViewController: targetViewController)
             return true
         }